HTTPS

+++ routes/dishRouter.js
@@ -21,6 +21,7 @@ dishRouter.route('/')
     .catch((err) => next(err));
 })
 .post(authenticate.verifyUser, (req, res, next) => {
+    authenticate.verifyAdmin(req.user.admin, res, next);
     Dishes.create(req.body)
     .then((dish) => {
         console.log('Dish Created ', dish);
@@ -31,10 +32,12 @@ dishRouter.route('/')
     .catch((err) => next(err));
 })
 .put(authenticate.verifyUser, (req, res, next) => {
+    authenticate.verifyAdmin(req.user.admin, res, next);
     res.statusCode = 403;
     res.end('PUT operation not supported on /dishes');
 })
 .delete(authenticate.verifyUser,(req, res, next) => {
+    authenticate.verifyAdmin(req.user.admin, res, next);
     Dishes.remove({})
     .then((resp) => {
         res.statusCode = 200;
@@ -56,10 +59,12 @@ dishRouter.route('/:dishId')
     .catch((err) => next(err));
 })
 .post(authenticate.verifyUser,(req, res, next) => {
+    authenticate.verifyAdmin(req.user.admin, res, next);
     res.statusCode = 403;
     res.end('POST operation not supported on /dishes/'+ req.params.dishId);
 })
 .put(authenticate.verifyUser,(req, res, next) => {
+    authenticate.verifyAdmin(req.user.admin, res, next);
     Dishes.findByIdAndUpdate(req.params.dishId, {
         $set: req.body
     }, { new: true })
@@ -71,6 +76,7 @@ dishRouter.route('/:dishId')
     .catch((err) => next(err));
 })
 .delete(authenticate.verifyUser,(req, res, next) => {
+    authenticate.verifyAdmin(req.user.admin, res, next);
     Dishes.findByIdAndRemove(req.params.dishId)
     .then((resp) => {
         res.statusCode = 200;
@@ -129,6 +135,7 @@ dishRouter.route('/:dishId/comments')
         + req.params.dishId + '/comments');
 })
 .delete(authenticate.verifyUser,(req, res, next) => {
+    authenticate.verifyAdmin(req.user.admin, res, next);
     Dishes.findById(req.params.dishId)
     .then((dish) => {
         if (dish != null) {
@@ -182,7 +189,8 @@ dishRouter.route('/:dishId/comments/:commentId')
 .put(cors.corsWithOptions, authenticate.verifyUser, (req, res, next) => {
     Dishes.findById(req.params.dishId)
     .then((dish) => {
-        if (dish != null && dish.comments.id(req.params.commentId) != null) {
+        if (dish != null && dish.comments.id(req.params.commentId) != null 
+        && dish.comments.id(req.params.commentId).author.equals(req.user._id)) {
             if (req.body.rating) {
                 dish.comments.id(req.params.commentId).rating = req.body.rating;
             }
@@ -216,7 +224,8 @@ dishRouter.route('/:dishId/comments/:commentId')
 .delete(cors.corsWithOptions, authenticate.verifyUser, (req, res, next) => {
     Dishes.findById(req.params.dishId)
     .then((dish) => {
-        if (dish != null && dish.comments.id(req.params.commentId) != null) {
+        if (dish != null && dish.comments.id(req.params.commentId) != null
+        && dish.comments.id(req.params.commentId).author.equals(req.user._id)) {
 
             dish.comments.id(req.params.commentId).remove();
             dish.save()
